I hate adding lines to the program....
The menus needed their own callbacks because the menu callback is different
than the GtkItemFactoryEntry callback. And because I was using the
user_data to determine which feed was the target, that wasn't being passed
correctly from the real menu. Anyway.... I'm not sure if this code is the
best, although it is very simple. I want to redo the menu anyway. I find
it very hard to navigate, and it is one of the first things that I didn't
like with the interface... Hmm. This is a long changelog entry. My roomate,
Jonathan Champ wanted his name in a changelog entry. He has helped me with
debugging.