I accidently left some printf lines in the font code from the last post.
Also, I made the feed cache saving on an as-needed basis.

It isn't quite correct, but I was interrupted and I forget what my
objection was. I wish that I could remember it. Oh yeah.... The feed
interval and title in the cache file are not re-saved after they are
changed. I think that this is OK (and that they should be removed) because
that information is duplicated in the feedlist.