The original intention of exporting all the GoogleSource edits to disk if you couldn't send them to Google were two fold:
1. Net requests were slow, so it took lots of time if you had many edits, and
   you might potential have closed Liferea if you didn't have the time.
2. Be able to store edits if you were offline while reading GoogleSource
   feeds, and then later closed Liferea.

(1) was fixed in an earlier commit today. (2) is rare, if you're using
a Laptop while offline, you would most probably suspend it rather than
turn it off.

This piece of code was too complex, and runs rarely, and so was impossible to
maintain.

(Additional TODO: there's a repetition in the logic of the GoogleSourceActionQueue with Liferea's update queue.)