PrefDialog: properly destroy the dialog
Before this commit, closing the dialog with Alt+F4 or through the
window manager (not through the close button which hides the dialog)
destroyed the GtkWidget, but the PreferencesDialog
object wasn't destroyed. Then when the preferences were opened
again, we tried to reuse the old GtkWidget because the
PreferencesDialog was still there, but failed.

We now destroy the PreferencesDialog object when closing the
dialog that way (the close button still hides the dialog).

In the future we should make PreferencesDialog (and other similar
widgets) a subclass of the appropriate GTK+ widget.

https://sourceforge.net/p/liferea/bugs/1063/