Fix incrementing bug and NULL bug in common_str_is_empty
As noted by Emilio Pozuelo Monfort, this function will hang in an
infinite loop if given input that starts with whitespace.  Forgot to
increment the pointer and to thoroughly test the function.

Also, if we say in documentation that the function returns FALSE on
NULL input, we can't use g_return_val_if_fail since that logs a
critical message and should only be used for catching programming
errors.

Signed-off-by: Lars Windolf <lars.lindner@gmail.com>