Stop calling Atom person constructs w/ URI invalid
The patched conditional is supposed to do the same thing as the one in
the block that parses `email` elements.

Namely, the `email` variable is initialised to `NULL`, so if it already
had an allocation when an `<email>` element is encountered, then there
must have been another `<email>` element in the person construct before,
and this second `<email>` constitutes an error per RFC 4287.

RFC 4287 specifies the same arity for the `<uri>` element, i.e. zero or
one `<uri>` elements. The code also initialises the `uri` variable to
`NULL` just like the `email` variable.

But the block that parses `<uri>` elements checks if the `uri` variable
is **not** `NULL`. This check will *always* be true when the *first*
`<uri>` element is encountered, so *any* person construct that contains
a URI will be declared invalid. This is a bug. The condition needs to be
negated.