Adds app menu and header menu to the ui file
The header bar isn't meant to be used with a menu bar. It is
recommended to have an app menu instead, which is only one menu,
and a header menu containing options related to the view.

Actions that are available in popup menus and navigation buttons aren't
duplicated in the app menu.