Adds header bar when prefered by environment
The toolbar is embedded in the header bar as an easy way to have
both the old style or the header bar. But the embedded toolbar may
not look very good on all Gtk themes ...

The header bar is used only when gtk_application_prefers_app_menu is
true (this is determined by the desktop environment).