GIthub #339: W32: Compatibility for time functions (pach by LRN)
gmtime_r and localtime_r are easy - just use gmtime_s and localtime_s,
which are basically the same, just with different names and arguments
swapped (thanks, Microsoft!).

However, strptime is not available in any way. The code added here
was borrowed from plibc, which borrowed it from R, which likely borrowed
it from somewhere else. It probably works.

Signed-off-by: Lars Windolf <lars.windolf@gmx.de>