Fix proxy settings
Both proxy settings "Auto Detect" and "No Proxy" result in a NULL
proxy hostname, which result in the default proxy being used at the
start with both options, and no proxy being used after changing to
either option.

This modifies functions to pass the proxy setting mode along with other
proxy settings and to set the SoupSession "proxy-uri" or
"proxy-resolver" properties depending on the setting mode, for both
network and webkit.

Requires libsoup >= 2.42 for the "proxy-resolver" property.