Replacing deprecated elements in prefs.ui
I changed target Gtk version to 3.14 (Debian stable version).

GtkHBox, GtkVBox and GtkTable are deprecated, replaced by GtkGrid.
GtkAlignment is deprecated, replaced by setting GtkWidget alignment and
margin properties.

The "margin-left" and "margin-right" GtkWidget properties are deprecated
and should be replaced by "margin-start" and "margin-end" but Glade uses
the old properties https://bugzilla.gnome.org/show_bug.cgi?id=722514 . I
replaced Glade's "margin_left" by "margin_start". The property is not
visible in Glade but it isn't removed when editing the file.

Glade marks the horizontal alignment property of labels as deprecated,
but it corresponds to the "xalign" property, which is a deprecated
property of GtkMisc since 3.14 and a new property of GtkLabel since
3.16. So I left them.

The "xalign" property of (check) buttons used to align the text in their
child widget (label) is deprecated. The doc suggest "access the child
widget directly if you need to control its alignment". I used instead
the "halign" GtkWidget property, which when set to "start" make the
whole button and its text align to the left in its allocated space.