Fixes a navigation history bug
Sometimes the "file:///" url could be added to history. When clicking on
a link in an item `liferea_html_view_handle_URL` is called and changes
`htmlview->priv->internal` to FALSE. WebKit's "uri" property changes to
the clicked url. If WebKit can't display it, it is passed to the
download tool. WebKit stays on the generated item html and the uri
property becomes "file:///" (the base url given to WebKit for generated
html). Because we are supposed to have navigated to an external url, it
is added to history by `liferea_htmlview_location_changed`. Going back
actually loads that uri and displays the filesystem.

With those changes `liferea_htmlview_location_changed` is called only on
WEBKIT_LOAD_COMMITTED : WebKit is loading the page corresponding to the
"uri" property. This also avoids adding to the history an url that
redirects to another.

`htmlview->priv->internal` is changed to FALSE in
`liferea_htmlview_location_changed` when WebKit is actually loading a
non locally generated URL.

The call to `browser_history_add_location` is removed from
`liferea_htmlview_launch_URL_internal` to avoid adding to history urls
that WebKit won't actually navigate to, or urls that redirect to others.

Also removes a cast warning.