Automatic update of themed icons #251
I renamed Liferea icons that are fallback for theme icons. I noticed
there is no online.png offline.png.

The pixmaps directory is appended to the GtkIconTheme search path.
Images directly in that directory will be used as ultimate fallback. We
don't actually *need* to add them to the hicolor theme but ["This is
legacy feature, and new icons should be put into the fallback icon
theme, which is called hicolor, rather than directly on the icon
path."][1] But for the future  rather than using the hicolor theme there
is [`gtk_icon_theme_add_resource_path`][2] which appear to do exactly
what we want. It requires using the GResource system.

I replaced use of GdkPixbuf for icons by GIcon. GtkCellRendererPixbuf
can use GIcons and the named (theme) icons are automatically updated on
theme change. GIcon is an interface and GdkPixbuf implements GIcon. The
GdkPixbuf implementation is used for favicons. The GThemedIcon
implementation is used for named icons.

Feeds without a favicon use a themed icon in its place in the views. For
this reason, I used the "gicon" property of the renderer. A GIcon
doesn't have a size, so I'm setting the larger size for wide view with
the renderer. Unless there is another property I'm not seeing, it
requires a GtkIconSize, which is an enum, not an arbitrary size so I
picked "drag and drop" (32px) and "small toolbar" (16px) sizes.

I removed largeIcon from the node as the renderer will take care of the
resizing. The kept icon is 32px, but we could also not scale it when
loading it from the cache.

[1]: https://developer.gnome.org/gtk3/stable/GtkIconTheme.html#gtk-icon-theme-set-search-path
[2]: https://developer.gnome.org/gtk3/stable/GtkIconTheme.html#gtk-icon-theme-add-resource-path