Check if selection really changed in callback
The GtkTreeSelection "changed" signal can be emitted even when the
selection didn't actually change.
By checking if the selected item really changed before calling
itemlist_selection_changed we avoid entering the loop in #208 : when
viewing a folder, with an item selected, if an update removes some
items the selection "changed" signal is emitted. This caused
itemlist_check_for_deferred_action to be called, removing the actually
selected item and moving the selection, marking many items as read in
the process.