Fix changing view layouts
In item_list_view.c, the GtkTreeModel is owned by the GtkTreeView, and
functions assume the GtkTreeView always exists. This modifies
item_list_view.c and itemview.c to make the ItemListView own the widget.
itemListViewContainer is also removed from ItemView : ItemView shouldn't
need to know the composition of the widget.

Some functions checked the ItemView mode and assumed that if it isn't
`ITEMVIEW_ALL_ITEMS` the ItemListView exists, but `itemview_set_mode`
doesn't change the widget layout, so the mode doesn't guaranty existence
of itemListView. Those functions are modified to check if itemListView
is NULL instead.

The `itemview_remove_item` and `itemview_update_item` functions relied
on ItemListView to know if an item is in the displayed item list. This
adds a `htmlview_contains_id` function to check if an item is present in
combined view.

The itemListView is also used to find the next unread item. In combined
view we currently don't keep track of read items, so this function
returns NULL if itemListView is NULL. It doesn't really matter because
`itemlist_select_next_unread` handles combined view feeds differently
already.