Workaround to avoid GtkPaned shrinking
Workaround for Gtk+ bug :
https://bugzilla.gnome.org/show_bug.cgi?id=733638

When trying to restore the position of the GtkPaned separator with both
children set to resize and a maximized window, the GtkPaned
automatically modifies its position after it is set.

To avoid this the resize property of the first child of the wide view
and normal view GtkPaned is set to TRUE only if Liferea isn't started
maximized. The LAST_WINDOW_MAXIMIZED setting is used for this because
`gtk_window_is_maximized` doesn't return a correct value at start (due
to window manager operations being asynchronous).

If the user unmaximize the window both children are set to resize, so
that they resize with the window. It is not necessary to disable resize
again when maximizing the window as the problem only occur when setting
the position.

The vertical separator in wide view still doesn't completely respect
the given position ...