Read default update interval from Cache-Control: max-age header
The Cache-Control: max-age header field contains the number of
seconds a response  is considered fresh. This suggests a default
update interval similar to syn:* and channel/ttl.

This implementation looks at syn, ttl, and maxage (minus any Age
header); and picks whichever is greatest as the update interval.

A new 5 minute minimum update interval is enforced as some feeds
have 0 or 60 seconds maxage.

https://tools.ietf.org/html/rfc7234#section-5.2.2.8
https://tools.ietf.org/html/rfc7234#section-4.2.3