Changes item popup menu to use GActions
Modifies the ui_popup_item_menu function to build the menu with actions,
and to apply to the clicked item.
Changes the button press event handler to not change selection on right
click.

The item passed to `ui_popup_item_menu` will be destroyed after the
function returns, it can't be directly passed to callbacks. Some actions
use only a link and are identical to actions in the web view context
menu. Those actions take the link as parameter and are added to the
application to be accessible to both menus, this also makes them
directly usable by plugins :
- open-link-in-tab
- open-link-browser
- open-link-in-external-browser
- social-bookmark-link
- copy-link-to-clipboard

Some actions require the item. In that case the actions take the
item->id as a parameter. Those actions are added directly to the popup
menu. The same callbacks are used for those actions and the actions that
use the selected item.
- copy-item-to-newsbin
- toggle-item-flag
- toggle-item-status
- remove-item