Avoid icons in main menu
Since refactoring to GtkBuilder and GAction the
main menu suddenly has menu icons again. This is
inconsistent with all other context menus and doesn't
follow HIG. Also the menu entries with icons are
not aligned with those without icons.

I'd prefer the icons to be hidden by GNOME defaults
so other desktop environments can use them, but I have
no idea on how to achieve it...