Fixes #245, #246: unread count of vfolder
The workaround for #280 in 8923577a9573c74c3a412fab1dba01c6626b4434 was
causing an infinite loop in `feedlist_unread_scan` (the assert that
should have stopped it doesn't work because `scanState` is equal to
`UNREAD_SCAN_FOUND_SELECTED` at that point).

This is a different workaround to keep "Mark all as read" enabled for
vfolders.