Avoid changing item state directly
When merging received items with the existing list, if the item state
has changed in the remote source, the new state is directly assigned to
the existing item.

This replaces the direct assignment with calls to
`item_read_state_changed`, `item_flag_state_changed` which performs
related operations like updating the unread count.