Embed YouTube videos
Extend xslt to render YouTube's embeded code if source is a YouTube video.

Another option would be to check the YT namespace, but adding another parser
might be overkill. Also, I was not able to find the YT schema.

Privacy wise, there would be a request everytime a feed is rendered, but this
should only happen if the feed is a YT MediaRSS feed or if the feed has a YT
link as main link. If description has a YT url, the embed code won't show.

Signed-off-by: Mikel Olasagasti Uranga <mikel@olasagasti.info>