Fixes #865: Fixes view mode switch
The way it works : `on_view_activate` is the callback to the stateful
action "set-view-mode".
Changing "set-view-mode"'s state calls this function.
The callback calls on `itemlist_set_view_mode` which calls
`itemview_set_layout`, etc ...

To keep things in sync, the view mode should only be set through
changing the action state. I restore the `liferea_shell_set_view_mode`
function in liferea_shell.c. It is a convenience function to get the
action and change it's state : this triggers the callback and the
view mode change.

I make `itemlist_set_view_mode` static so that it won't be used
directly. I replace it with `liferea_shell_set_view_mode` in feedlist.c.