Execute feed commands asynchronously
Currently, running a command to generate feeds is a blocking operation
that freezes the interface and makes the entire program irresponsive.
This commit makes these commands run asynchronously in background
allowing interaction with the program while they are running.

An arbitrary 30s timeout is imposed to child processes to prevent slow
or blocked commands from running forever.