Use liferea_shell_show_window when reactivating
Instead of calling gtk_window_deiconify and gtk_window_present directly,
use the wrapper liferea_shell_show_window which also moves the window to
the current desktop and restores the position if the window is hidden.

Note that we no longer restore position if iconified. I am not sure why
this was ever the case but it seems unnecessary. It also triggers a bug
on Cinnamon, whereas letting the window manager handle this seems to
work fine.