Add feature to save contents for a feed or vfolder to a file
This new feature allows saving/exporting the items from feeds, search
folders and news bins to a RSS 2.0 file. Typical use cases are an user
wanting to process feeds with an external tool or publish a static
feed after using Liferea to download, sort, or select the data.

Code and behavior are the same for all supported node types, allowing
users to, e.g., select the items manually (copying them to a news bin)
or automatically (using a search folder). It is also possible to export
cached items from standard feed subscriptions.

The only output format currently supported is RSS 2.0 but new ones
can be added if necessary.