Browse to requested URI on JS window creation
Some sites intercept hyperlink clicks in order to inject tracking
scripts or similar. They may then open the link in a new window via
JavaScript. In Liferea, this results in a new empty tab opening with no
content.

Browse to the request URI when one is provided. Opening a new window
without a URI in Liferea won't have much effect, but it could happen, so
that's handled as well.