Preserve browsing mode when marking items as read
Setting the selected item to NULL when marking items as read results in
odd behavior, since the content is still shown to the user but the item
list shows no item as selected.

Reverting this change and bailing when the browsing flag is set prevents
the browser content from being replaced with the feed content when
marking items as read.

The addition of itemview_select_item (NULL) to feedlist_mark_all_read is
attributed to #208, but I wasn't able to reproduce that issue after
making this change.