Free previous stylesheet and only reload on change
There was a memory leak in the previous implementation any time
render_get_css was called (which as it turns out can be often.)

Also, there's no need to reload the stylesheet if it hasn't changed.
Unfortunately there's no way to get the stylesheet back from WebKit, so
we keep a copy of it.