Write content to output.html with --debug-html
With the changes for reader mode, output.html no longer contains the
content of the item view. Write content to output.html in
liferea_browser_update to fix.

Note that this will not show changes made via JavaScript (e.g. Reader
Mode). The inspector can be used to view the page source in such cases.

Fixes #1092