Return more descriptive status codes from failed feed commands
Currently, when a feed command fails for any reason, the parser only
receives a fake 404 error. This does enough to signal the failed command to
Liferea but, since the status ends up in the UI with a descriptive message,
we can expose a bit more information on the cause of the failure by
hijacking a few extra HTTP status codes.

The code will now use:

 - 200 for all successful runs;
 - 404 if failed to run the command (file not found or no exec bit);
 - 504 if the command timed out;
 - 500 for all other failures (killed, non-zero return, etc.).

This is a relative improvement only, I would really like to replace it by
an elegant approach that does not depend on faking HTTP status codes for
an input that is not received from HTTP.