conf: fix crash if DE does not have a color scheme preference
Liferea crashed when running in a desktop environment that does not
have a color scheme preference. This is critical as this feature is
pretty new across DEs e.g. the "color-scheme" configuration key was
only added to Gnome in 19/09/2021 (commit
e3c2c421d78a8bfff466ad2180781cbaacce0319 from gsettings-desktop-schemas
repo) and only released in Gnome 42. Now we check if the configuration
key exists before trying to use it.