Add a "flat view" mode for the feed list
Idea is to show all feeds in a single list, without folders and other
grouping elements, acting as a alternate for the reduced view mode.
Persisting the choice for this mode also requires adding new config
keys to the gsettings schema.

This is the first part of a two-commit change (actually split after a
'merge --squash') that will also add a filter box for feed titles.