Set prgname to match the reverse dns style application id
When launched on wayland with `WAYLAND_DEBUG=1` Liferea sets the app id
"liferea" but not net.sourceforge.liferea:

```
[1225601.077]  -> xdg_toplevel@41.set_app_id("liferea")
```

When launched on KDE Wayland session, it expects the desktop file name
to match the application id to look up the correct app icon to show in
taskbar app swicther etc. But instead it tries too look up an icon
called "liferea" which does not exist. So liferea shows up with the
default wayland icon instead of the app icon.

Setting WM_CLASS here won't work and we can't export two desktop files.

This does not effect GNOME because it tries harder to guess the app icon